Fixed PV user profile to take into account the panel count. Refactored d3a_path to be part of d3a.util module and not of the predefined PV.
