+++ src/d3a/setup/default_addcurve.py
@@ -8,7 +8,7 @@ from d3a.models.strategy.load_hours_fb import LoadHoursStrategy
 from d3a.models.appliance.pv import PVAppliance
 from d3a.models.strategy.pv import PVStrategy
 from d3a.models.strategy.predefined_load import DefinedLoadStrategy
-from d3a.models.strategy.predefined_pv import d3a_path
+from d3a.util import d3a_path
 import os
