+++ src/d3a/setup/default_csv.py
@@ -2,7 +2,7 @@
 from d3a.models.appliance.switchable import SwitchableAppliance
 from d3a.models.area import Area
 # from d3a.models.strategy.commercial_producer import CommercialStrategy # NOQA
-from d3a.models.strategy.predefined_pv import d3a_path
+from d3a.util import d3a_path
 from d3a.models.strategy.storage import StorageStrategy
 from d3a.models.strategy.load_hours_fb import LoadHoursStrategy, CellTowerLoadHoursStrategy
 from d3a.models.strategy.predefined_load import DefinedLoadStrategy
