+++ src/d3a/setup/sam_config_small.py
@@ -5,7 +5,7 @@ from d3a.models.strategy.storage import StorageStrategy
 from d3a.models.appliance.pv import PVAppliance
 from d3a.models.strategy.pv import PVStrategy
 from d3a.models.strategy.predefined_load import DefinedLoadStrategy
-from d3a.models.strategy.predefined_pv import d3a_path
+from d3a.util import d3a_path
 import os
