+++ src/d3a/setup/strategy_tests/user_profile_pv_csv.py
@@ -9,7 +9,7 @@ from d3a.models.area import Area
 from d3a.models.strategy.predefined_pv import PVUserProfileStrategy
 # from d3a.models.strategy.storage import StorageStrategy
 from d3a.models.strategy.load_hours_fb import LoadHoursStrategy
-from d3a.models.strategy.predefined_pv import d3a_path
+from d3a.util import d3a_path
 import os
