+++ tests/test_settings_file.py
@@ -2,7 +2,7 @@ import os
 import unittest
 
 from d3a.util import IntervalType
-from d3a.models.strategy.predefined_pv import d3a_path
+from d3a.util import d3a_path
 from d3a.models.strategy.const import ConstSettings
 from d3a.util import read_settings_from_file
 from d3a.util import update_advanced_settings
